% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_fgb_to_gdb.R
\name{convert_fgb_to_gdb}
\alias{convert_fgb_to_gdb}
\title{Convert FlatGeoBuf (FGB) Files to Geodatabase (GDB) Layers}
\usage{
convert_fgb_to_gdb(
  input_folder,
  gdb_path,
  overwrite = FALSE,
  crs = "EPSG:3857"
)
}
\arguments{
\item{input_folder}{Character. Path to the folder containing FGB files.}

\item{gdb_path}{Character. File path to the output GDB file.}

\item{overwrite}{Logical. Whether to overwrite existing layers in the GDB. Default is FALSE.}

\item{crs}{Character. Coordinate Reference System to transform the data. Default is "EPSG:3857".}
}
\description{
This function converts all FlatGeoBuf (.FGB) files within a specified folder into layers of a specified Geodatabase (GDB) file.
}
\examples{
\dontrun{
convert_fgb_to_gdb("/path/to/fgb_folder", "/path/to/output.gdb", overwrite = FALSE, crs = "EPSG:3857")
}
}
