% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_html_table.R
\name{extract_html_table}
\alias{extract_html_table}
\title{Extract table following a specific anchor tag by name attribute}
\usage{
extract_html_table(html_file, anchor_id)
}
\arguments{
\item{html_file}{A character string specifying the path to the HTML file.}

\item{anchor_id}{A character string specifying the \code{name} attribute of the anchor (\verb{<a>}) tag to locate.}
}
\value{
A data frame containing the extracted table.
}
\description{
This function extracts the summary and trip tables from Warren TDM HTML file.
}
\examples{
\dontrun{
summary_table <- extract_html_table("Model.html", "id6")
# View the extracted table
View(summary_table[[1]])

trip_table <- extract_html_table("Model.html", "id52")
# View the extracted table
View(trip_table[[1]])
}

}
