% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ugrc.R
\name{load_ugrc_data}
\alias{load_ugrc_data}
\title{Load Dataset from UGRC SGID}
\usage{
load_ugrc_data(table_name, as_sf = TRUE)
}
\arguments{
\item{table_name}{A character string specifying the name of the table to retrieve (e.g., \code{"schema.table_name"}).}

\item{as_sf}{Logical. If \code{TRUE} (default), the data is returned as an \code{sf} object. If \code{FALSE}, the geometry column is dropped and a regular data frame is returned.}
}
\value{
An \code{sf} object or a data frame containing the data from the specified table.
}
\description{
Load Dataset from UGRC SGID
}
\details{
Loads a dataset from the UGRC SGID database. By default, the dataset is returned as an \code{sf} object for spatial data analysis.
}
\examples{
\dontrun{
# Load UTA Transit Lines
uta_lines <- load_ugrc_data("transportation.uta_routes_and_ridership")

# Load UTA Transit Stops
uta_stops <- load_ugrc_data("transportation.uta_stops_and_ridership")
}

}
