% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ugrc.R
\name{load_ugrc_vars}
\alias{load_ugrc_vars}
\title{Load List of Datasets from UGRC SGID}
\usage{
load_ugrc_vars()
}
\value{
A data frame containing two columns: \code{table_schema} and \code{table_name}, listing the schemas and tables in the SGID database.
}
\description{
Load List of Datasets from UGRC SGID
}
\details{
Retrieves a list of available datasets in the UGRC SGID database, including their schema and table names.
}
\examples{
# Check list of available datasets
ugrc_vars <- load_ugrc_vars()
View(ugrc_vars)

}
