% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_aggregate.R
\name{st_aggregate}
\alias{st_aggregate}
\title{Aggregate sf objects}
\usage{
st_aggregate(x, by, var, fun)
}
\arguments{
\item{x}{sf object}

\item{by}{name of the variable of grouping elements}

\item{var}{name(s) of the variable(s) to aggregate}

\item{fun}{function(s) to compute the summary statistics}
}
\value{
An sf object is returned
}
\description{
Aggregate sf objects
}
\details{
Geometries and attributes are aggregated.
Source: https://gist.github.com/rCarto/bb47aff0a02e808d2bf64f2d8c5db7d8#file-st_aggregate-r
}
\examples{
\dontrun{
library(sf)
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
nc$dummy <- "ZONE_A"
nc$dummy[25:50] <- "ZONE_B"
nc$dummy[51:100] <- "ZONE_C"
r <- st_aggregate(nc, "dummy", c("BIR74", "NWBIR74"), c("mean", "median"))
plot(nc)
plot(r)
}

}
